Components: Move SiteIcon to Blocks
