+++ assets/stylesheets/_components.scss
@@ -136,7 +136,7 @@
 @import 'components/seo/preview-upgrade-nudge/style';
 @import 'components/signup-form/style';
 @import 'components/signup-site-title/style';
-@import 'components/site-icon/style';
+@import 'blocks/site-icon/style';
 @import 'components/site-selector/style';
 @import 'components/site-selector-modal/style';
 @import 'components/site-title-example/style';
