+++ client/blocks/site-icon/index.jsx
