+++ client/blocks/site-icon/style.scss
