+++ client/my-sites/pages/page.jsx
@@ -17,7 +17,7 @@ var updatePostStatus = require( 'lib/mixins/update-post-status' ),
 	Gridicon = require( 'components/gridicon' ),
 	PopoverMenu = require( 'components/popover/menu' ),
 	PopoverMenuItem = require( 'components/popover/menu-item' ),
-	SiteIcon = require( 'components/site-icon' ),
+	SiteIcon = require( 'blocks/site-icon' ),
 	helpers = require( './helpers' ),
 	analytics = require( 'lib/analytics' ),
 	utils = require( 'lib/posts/utils' ),
