+++ client/my-sites/sidebar-navigation/sidebar-navigation.jsx
@@ -8,7 +8,7 @@ var React = require( 'react' );
  */
 var SidebarNavigation = require( 'components/sidebar-navigation' ),
 	AllSitesIcon = require( 'my-sites/all-sites-icon' ),
-	SiteIcon = require( 'components/site-icon' ),
+	SiteIcon = require( 'blocks/site-icon' ),
 	sites = require( 'lib/sites-list' )();
 
 module.exports = React.createClass( {
