+++ client/reader/following-edit/placeholder.jsx
@@ -4,7 +4,7 @@
 var React = require( 'react' );
 
 var Card = require( 'components/card' );
-var SiteIcon = require( 'components/site-icon' );
+var SiteIcon = require( 'blocks/site-icon' );
 
 module.exports = React.createClass( {
